Authorization

+++ user-info-service/src/main/java/com/predictmatch/userinfo/service/UserInfoServiceImpl.java
@@ -1,10 +1,18 @@
 package com.predictmatch.userinfo.service;
 
+import com.auth0.jwt.interfaces.DecodedJWT;
 import com.predictmatch.userinfo.dao.UserInfo;
-import com.predictmatch.userinfo.dto.*;
+import com.predictmatch.userinfo.dao.auth.User;
+import com.predictmatch.userinfo.dto.TeamDto;
+import com.predictmatch.userinfo.dto.TeamRequestDto;
+import com.predictmatch.userinfo.dto.UserInfoRequest;
+import com.predictmatch.userinfo.dto.UserInfoResponse;
+import com.predictmatch.userinfo.dto.auth.RegisterRequest;
 import com.predictmatch.userinfo.exceptions.UserAlreadyExistsException;
 import com.predictmatch.userinfo.mapper.Mapper;
 import com.predictmatch.userinfo.repository.UserInfoRepository;
+import com.predictmatch.userinfo.security.JwtUtils;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +23,7 @@ import javax.transaction.Transactional;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class UserInfoServiceImpl implements UserInfoService {
 
     @Autowired
@@ -25,6 +34,9 @@ public class UserInfoServiceImpl implements UserInfoService {
 
     @Autowired
     ValidationService validationService;
+
+    @Autowired
+    JwtUtils jwtUtils;
 
     @Transactional
     @Override
@@ -55,35 +67,53 @@ public class UserInfoServiceImpl implements UserInfoService {
         return ResponseEntity.ok( Mapper.userInfoEntityToDto( storedUser.get(),team ));
 
     }
+
     @Transactional
     @Override
-    public ResponseEntity<UserInfoResponse> createUser(UserInfoRequest request) {
+    public UserInfoResponse createUserProfile(RegisterRequest request, User user) {
 
         if(validationService.checkIfUserAlreadyExists(request.getUsername().trim())) {
             throw new UserAlreadyExistsException( request.getUsername());
         }
 
-       UserInfo user = Mapper.userInfoRequestToEntity( request );
-
-       userInfoRepository.save( user );
+       UserInfo userInfo = Mapper.registerRequestToUserInfoEntity( request,user );
+        log.info("User profile is creating...");
+       userInfoRepository.save( userInfo );
 
        Optional<TeamDto> team = Optional.ofNullable( teamService.findTeam( request.getTeamId() ));
        TeamDto userTeam = null;
 
-       if(team.isPresent())
+       if(team.isPresent()) {
            userTeam=team.get();
+           log.info( "Team {} added to user profile",team.get().getName() );
+       } else {
+           log.error( "Currently cannot find team with id: "+request.getTeamId());
+       }
 
-       return ResponseEntity.status( HttpStatus.CREATED ).body( Mapper.userInfoEntityToDto( user,userTeam));
+
+       return Mapper.userInfoEntityToDto( userInfo,userTeam);
 
     }
+
     @Transactional
     @Override
-    public ResponseEntity<UserInfoResponse> changeFavouriteTeam(Long id, TeamRequestDto teamRequest) {
+    public ResponseEntity<?> changeFavouriteTeam(String username, TeamRequestDto teamRequest, String token) {
 
-        Optional<UserInfo> storedUser = userInfoRepository.findById( id );
+        DecodedJWT decodedToken = jwtUtils.decodeJwtToken( token );
+
+        long tokenUserId = Long.parseLong(decodedToken.getClaim( "user_id").toString());
+
+        String tokenUsername = decodedToken.getSubject();
+
+
+
+        if(!tokenUsername.equals( username ))
+            return ResponseEntity.status( HttpStatus.UNAUTHORIZED.value()).body( "Error: Unauthorized operation!");
+
+        Optional<UserInfo> storedUser = userInfoRepository.findById( tokenUserId );
 
         if(storedUser.isEmpty())
-            throw new EntityNotFoundException("Not found user with id: "+id);
+            throw new EntityNotFoundException("Not found user with id: "+tokenUserId);
 
         UserInfo user = storedUser.get();
 
@@ -97,12 +127,22 @@ public class UserInfoServiceImpl implements UserInfoService {
     }
     @Transactional
     @Override
-    public ResponseEntity<UserInfoResponse> updateUserInfo(Long id, UserInfoRequest request) {
+    public ResponseEntity<?> updateUserInfo(String username, UserInfoRequest request, String token) {
 
-        Optional<UserInfo> storedUser = userInfoRepository.findById( id );
+        DecodedJWT decodedToken = jwtUtils.decodeJwtToken( token );
+
+        long tokenUserId = Long.parseLong(decodedToken.getClaim( "user_id").toString());
+
+        String tokenUsername = decodedToken.getSubject();
+
+        if(!tokenUsername.equals(username  ))
+            return ResponseEntity.status( HttpStatus.UNAUTHORIZED.value()).body( "Error: Unauthorized operation!");
+
+
+        Optional<UserInfo> storedUser = userInfoRepository.findById( tokenUserId );
 
         if(storedUser.isEmpty())
-            throw new EntityNotFoundException("Not found user with id: "+id);
+            throw new EntityNotFoundException("Not found user with id: "+tokenUserId);
 
         UserInfo user = storedUser.get();
 
@@ -140,13 +180,24 @@ public class UserInfoServiceImpl implements UserInfoService {
 
     @Transactional
     @Override
-    public ResponseEntity<String> removeUser(Long id) {
-        Optional<UserInfo> storedUser = userInfoRepository.findById( id );
+    public ResponseEntity<String> removeUser(String username, String token) {
+
+        DecodedJWT decodedToken = jwtUtils.decodeJwtToken( token );
+
+        long tokenUserId = Long.parseLong(decodedToken.getClaim( "user_id").toString());
+
+        String tokenUsername = decodedToken.getSubject();
+
+        if(!tokenUsername.equals(username  ))
+            return ResponseEntity.status( HttpStatus.UNAUTHORIZED.value()).body( "Error: Unauthorized operation!");
+
+
+        Optional<UserInfo> storedUser = userInfoRepository.findById( tokenUserId );
 
         if(storedUser.isEmpty())
-            return ResponseEntity.status( HttpStatus.NOT_FOUND).body( "User: "+id+" does not exist" );
+            return ResponseEntity.status( HttpStatus.NOT_FOUND).body( "User: "+username+" does not exist" );
 
         userInfoRepository.delete(  storedUser.get());
-        return ResponseEntity.status( HttpStatus.ACCEPTED).body( "User with id: "+id+" was removed" );
+        return ResponseEntity.status( HttpStatus.ACCEPTED).body( "User with id: "+username+" was removed" );
     }
 }
